requirements.txt
requests~=2.27.1
qrcode~=7.3.1
Pillow~=9.0.1
pyzbar~=0.1.8
python-telegram-bot~=13.11